Implemented 'quick fix' for floor traversal error
I was trying to remove the Hero from each floor.objectList and place him
back in the new floor he moved to. Was getting an error for continuing
to iterate through the list of objects after having deleted one element
from it (Hero). Moved the deletion code to after the collision loop.